pyproject.toml
[project]
name = "paperless-intelligence"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.27.0",
    "nicegui>=1.4.22",
    "pydantic>=2.7.0",
    "psycopg[binary]>=3.2.3",
    "python-dotenv>=1.0.1",
    "redis>=5.0.8",
]
